fix: Align ColorTable colors to the base16 spec
Below is the default mapping in Windows Console showing how colors are used
by ConsoleHost to display different elements.

| ANSI         | ColorTable | concfg key (PowerShell) | Properties Option  |
|--------------|------------|-------------------------|--------------------|
| White        | 07         | gray                    | Screen Text        |
| Black        | 00         | black                   | Screen Background  |
| Magenta      | 05         | dark_magenta            | Popup Text         |
| Bright White | 15         | white                   | Popup Background   |

* ColorTable[07] (concfg's `gray`, ANSI White) is used as the Screen Text
(ANSI foreground) color in ConsoleHost.
* ColorTable[15] (concfg's `white`, ANSI Bright White) is used as the Popup
Background color in COnsoleHost.

Before this change, `base07` (Light Background in base16 spec) was used for
`gray` and `base06` (Light Foreground in base16 spec) was used for `white`.
This resulted in a background base16 color set to a concfg key (`gray`) which
is used as a foreground color, and a foreground base16 color set to a concfg
key (`white`) which is used as a background color. Further, the colors used
in `screen_colors` and `popup_colors` were inverted in order to align to the
default mapping of ConsoleHost.

In this commit, `gray` is changed to use `base05` (Default Foreground in base16
spec), `white` is changed to use `base07` (Light Background in base16 spec).
With these changes, ColorTable colors now semantically aligns to base16 colors.
The inverted usage of colors in `screen_colors` and `popup_colors` is removed,
leaving a better alignment between ConsoleHost's ColorTable and base16.

Signed-off-by: Chawye Hsu <su+git@chawyehsu.com>